requirements.txt
requests>=2.28
psycopg2-binary>=2.9

yfinance>=0.2.65
pandas>=1.3.5,<2.0
PyMySQL>=1.0.2
